fix(ta_func): compute candle range in double inside TA_S_ bodies (#138)
TA_REALBODY/TA_UPPERSHADOW/TA_LOWERSHADOW/TA_HIGHLOWRANGE in ta_utility.h read
their input arrays with no cast. In a TA_S_CDL* body those arrays are const
float[], so TA_CANDLERANGE/TA_CANDLEAVERAGE compute the range in float and round
before it widens into the double accumulator -- breaking PR #33's "TA_S_ computes
in double throughout". The generator widens the longhand ta_realbody it inlines
but leaves the macro path float; only these two macros survive in generated code,
and only they escape the widening.

Fix: widen the subtraction operands to double in the four leaf macros, mirrored
into the TA_STREAM_* leaves. The >= tests need no cast (float->double is
order-preserving). Strict no-op for the double variants -- (double) on a double
-- so no shipped output moves: fuzz-064 stays bit-identical to v0.6.4 on all 61
CDL*. C-only: Java already inlines-and-widens per variant, Rust has no
single-precision variant, .NET P/Invokes C.

Latent, not a live wrong answer. Subtracting two floats within a factor of two is
exact (Sterbenz), so no realistic bar differs -- which is why #137's gate
reported CDL* green and its fuzz found 0/61. It bites only when one operand sits
below the other's float ULP, a high/low ratio past ~2^24.

Regression test (test_candle_precision -- the honest one #138 said was hard):
CDLDOJI at {HighLow,10,0.1}; bar 9 H=2^30 L=31 makes float(H-L) round up by 31,
bar 10 body tuned into the resulting 0.31-wide threshold gap. Pre-fix
TA_CDLDOJI=0 while TA_S_CDLDOJI=100; post-fix all four variants agree at 0. Runs
on a bare ./ta_regtest so the autotools dist nightly covers it.

Verified: bare ./ta_regtest (VARIANT gate 167 funcs x 31104 vectors bit-identical
+ new test green); fuzz-064 PASS, 0 failures.

+++ src/ta_func/ta_utility.h
@@ -189,10 +189,14 @@
  * Don't use i++ or func(i) with these macros !
  */
 
-#define TA_REALBODY(IDX)        ( std_fabs( inClose[IDX] - inOpen[IDX] ) )
-#define TA_UPPERSHADOW(IDX)     ( inHigh[IDX] - ( inClose[IDX] >= inOpen[IDX] ? inClose[IDX] : inOpen[IDX] ) )
-#define TA_LOWERSHADOW(IDX)     ( ( inClose[IDX] >= inOpen[IDX] ? inOpen[IDX] : inClose[IDX] ) - inLow[IDX] )
-#define TA_HIGHLOWRANGE(IDX)    ( inHigh[IDX] - inLow[IDX] )
+/* The subtraction operands are widened to double so these compute in double even
+ * when the arrays are const float[] (the TA_S_ bodies) — a no-op for the double
+ * variants, PR #33's "double throughout" for the single-precision ones (#138).
+ * The >= tests need no cast: float->double is order-preserving. */
+#define TA_REALBODY(IDX)        ( std_fabs( (double)inClose[IDX] - (double)inOpen[IDX] ) )
+#define TA_UPPERSHADOW(IDX)     ( (double)inHigh[IDX] - ( inClose[IDX] >= inOpen[IDX] ? (double)inClose[IDX] : (double)inOpen[IDX] ) )
+#define TA_LOWERSHADOW(IDX)     ( ( inClose[IDX] >= inOpen[IDX] ? (double)inOpen[IDX] : (double)inClose[IDX] ) - (double)inLow[IDX] )
+#define TA_HIGHLOWRANGE(IDX)    ( (double)inHigh[IDX] - (double)inLow[IDX] )
 #define TA_CANDLECOLOR(IDX)     ( inClose[IDX] >= inOpen[IDX] ? 1 : -1 )
 
 #define TA_CANDLERANGETYPE(SET) (TA_Globals->candleSettings[TA_##SET].rangeType)
@@ -221,10 +225,10 @@
  * arrays.  Any change to the batch macros must be mirrored here, or the
  * streams lose bit-exactness with batch on the affected range type.
  */
-#define TA_STREAM_REALBODY(O,H,L,C)     ( std_fabs( (C) - (O) ) )
-#define TA_STREAM_UPPERSHADOW(O,H,L,C)  ( (H) - ( (C) >= (O) ? (C) : (O) ) )
-#define TA_STREAM_LOWERSHADOW(O,H,L,C)  ( ( (C) >= (O) ? (O) : (C) ) - (L) )
-#define TA_STREAM_HIGHLOWRANGE(O,H,L,C) ( (H) - (L) )
+#define TA_STREAM_REALBODY(O,H,L,C)     ( std_fabs( (double)(C) - (double)(O) ) )
+#define TA_STREAM_UPPERSHADOW(O,H,L,C)  ( (double)(H) - ( (C) >= (O) ? (double)(C) : (double)(O) ) )
+#define TA_STREAM_LOWERSHADOW(O,H,L,C)  ( ( (C) >= (O) ? (double)(O) : (double)(C) ) - (double)(L) )
+#define TA_STREAM_HIGHLOWRANGE(O,H,L,C) ( (double)(H) - (double)(L) )
 #define TA_STREAM_CANDLERANGE(SET,O,H,L,C) \
     ( TA_CANDLERANGETYPE(SET) == TA_RangeType_RealBody ? TA_STREAM_REALBODY(O,H,L,C) : \
     ( TA_CANDLERANGETYPE(SET) == TA_RangeType_HighLow  ? TA_STREAM_HIGHLOWRANGE(O,H,L,C) : \
